Implemented ability to send manual report

+++ src/com/sibext/android/manager/CrashCatcherManager.java
@@ -37,61 +37,64 @@ import com.sibext.crashcatcher.R;
 public class CrashCatcherManager {
     private static final String TAG = "[CCL] CrashCatcherManager";
 
+    public static final String MANUAL_FLAG_KEY = "MANUAL_FLAG_KEY";
+    
     private Context context;
     private Class<?> catchClass = null;
 
     private void init() {
-        ApplicationInfo ai = null;
-        try {
-            ai = context.getPackageManager().getApplicationInfo(context.getPackageName(), PackageManager.GET_META_DATA);
-            if (ai != null) {
-                String className = ai.metaData.getString(context.getString(R.string.metadata_reporter_key));
-                if (null != className) {
-                    catchClass = Class.forName(className);
-                }
-            }
-        } catch (NameNotFoundException e) {
-            Log.e(TAG, "Can't get init params", e);
-        } catch (ClassNotFoundException e1) {
-            Log.w(TAG, "Using default reporter...");
-        }
+        catchClass = getCatchClass();
+
         Thread.setDefaultUncaughtExceptionHandler(new Thread.UncaughtExceptionHandler() {
             private volatile boolean alreadyCrashed = false;
             @Override
             public void uncaughtException(Thread paramThread, final Throwable e) {
                 if (alreadyCrashed) return;
                 alreadyCrashed = true;
-                try {
-                    final String stackTrace = StackTraceHelper.getStackTrace(e);
-                    Log.e(TAG, "Error: " + stackTrace);
-                    final Intent crashedIntent;
-                    if (catchClass != null) {
-                        crashedIntent = new Intent(context.getApplicationContext(), catchClass);
-                    } else {
-                        crashedIntent = new Intent(context.getApplicationContext(), getDefaultReporterClass());
-                    }
-                    crashedIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                    crashedIntent.putExtra(CatchActivity.TRACE_INFO, stackTrace);
-                    final PendingIntent pendingIntent = PendingIntent.getActivity(context, 0, crashedIntent, 0);
-                    ((Activity) context).runOnUiThread(new Runnable() {
-                        public void run() {
-                            try {
-                                pendingIntent.send(context.getApplicationContext(), 0, null);
-                                Log.d(TAG, "sent new crash");
-                                android.os.Process.killProcess(android.os.Process.myPid());
-                                System.exit(10);
-                            } catch (PendingIntent.CanceledException e) {
-                                Log.d(TAG, "Can't start activity", e);
-                            }
-                        }
-                    });
-                } catch (Throwable e1) {
+				try{
+					final String stackTrace = StackTraceHelper.getStackTrace(e);
+		            Log.e(TAG, "Error: " + stackTrace);
+		            sendReport(stackTrace, false);
+				} catch (Throwable e1) {
                     try {
                         Log.e(TAG, "Can't handle  the crash", e1);
                     } catch (Throwable e2) {}
                 }
             }
 
+        });
+    }
+    
+    public void manualSendReport(){
+        sendReport("Manual", true);
+    }
+    
+    public void sendReport(String stackTrace, final boolean manual){
+        final Intent crashedIntent;
+        if (catchClass != null) {
+            crashedIntent = new Intent(context.getApplicationContext(), catchClass);
+        } else {
+            crashedIntent = new Intent(context.getApplicationContext(), getDefaultReporterClass());
+        }
+        crashedIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+        crashedIntent.putExtra(CatchActivity.TRACE_INFO, stackTrace);
+        crashedIntent.putExtra(MANUAL_FLAG_KEY, manual);
+        final PendingIntent pendingIntent = PendingIntent.getActivity(context, 0, crashedIntent, 0);
+        ((Activity) context).runOnUiThread(new Runnable() {
+            public void run() {
+                try {
+                    pendingIntent.send(context.getApplicationContext(), 0, null);
+                    if(manual){
+                        Log.d(TAG, "sent manual report");
+                    } else {
+                        Log.d(TAG, "sent new crash");
+                         android.os.Process.killProcess(android.os.Process.myPid());
+                         System.exit(10);
+                    }
+                } catch (PendingIntent.CanceledException e) {
+                    Log.d(TAG, "Can't start activity", e);
+                }
+            }
         });
     }
